Added Reporting Scenes to view aggregated totals of Appointments by Month and Type as well as Customers by FLD. Schedules for each contact can be viewed in the Contact Schedule Scene.
Added lambda expressions of the Functional interface Filterer to simplify some currently used filters that were done using altered queries.

Added Javadoc Comments throughout

Project is complete for first draft

+++ src/main/java/controller/MainMenuController.java
@@ -6,11 +6,8 @@ import helper.Navigation;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.Label;
+import javafx.scene.control.*;
 import javafx.fxml.Initializable;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import model.Appointment;
 import model.Customer;
@@ -18,8 +15,15 @@ import model.Customer;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
+/**
+ * The Main Menu Controller
+ * This is the application's main dashboard.
+ * Displays tables for the list of Customers and the list of Appointmnets
+ * Many Navigation options on this page to access Add/Update/Delete functionality
+ */
 public class MainMenuController implements Initializable{
     @FXML
     private Label welcomeLabel;
@@ -67,38 +71,78 @@ public class MainMenuController implements Initializable{
     private TableColumn appointmentsUserIDColumn;
     public static Customer selectedCustomer;
     public static Appointment selectedAppointment;
-
+    /**
+     * Method that Filters the Appointments table to match the selected customer
+     * on SelectCustomer button click
+     * Incorporates lambda expression defined in the Query class to filter appointments
+     * @throws SQLException
+     */
     public void onSelectCustomer() throws SQLException {
         Customer selection = (Customer) customersTable.getSelectionModel().getSelectedItem();
 
         if (selection != null) {
-            appointmentsTable.setItems(AppointmentsQuery.getCustomerAppointments(selection.getCustomer_ID()));
+            appointmentsTable.setItems
+                    (AppointmentsQuery.customerFilter.extract(AppointmentsQuery.getAllAppointments(),
+                            selection.getCustomer_ID()));
             errorLabel.setText("");
         }
         else errorLabel.setText("No Customer Selected!");
     }
+
+    /**
+     * Radio Button selection that will show All appointmnets
+     * @throws SQLException
+     */
     public void onAllAppointmentsSelected() throws SQLException {
         appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
         errorLabel.setText("");
     }
+    /**
+     * Radio Button selection that shows all appointmnets in the current month
+     * @throws SQLException
+     */
     public void onMonthAppointmentsSelected() throws SQLException {
         appointmentsTable.setItems(AppointmentsQuery.getMonthAppointments());
         errorLabel.setText("");
     }
+    /**
+     * Radio button selection that shows all appoinments in the current week (Sunday -> Saturday)
+     * @throws SQLException
+     */
     public void onWeekAppointmentsSelected() throws SQLException {
         appointmentsTable.setItems(AppointmentsQuery.getWeekAppointments());
         errorLabel.setText("");
     }
+    /**
+     * Search functionality for the customer table to fund based on ID or Substring
+     * @throws SQLException
+     */
     public void onSearchCustomer() throws SQLException {
         String searchText = searchCustomerText.getText();
         customersTable.setItems(CustomersQuery.getCustomers(searchText));
     }
+    /**
+     * Switches to the AddCustomer scene
+     * @param e
+     * @throws IOException
+     */
     public void onAddCustomer(ActionEvent e) throws IOException {
         Navigation.switchToAddCustomer(e);
     }
+    /**
+     * Switches to Add Appointment scene
+     * @param e
+     * @throws IOException
+     */
     public void onAddAppointment(ActionEvent e) throws IOException {
         Navigation.switchToAddAppointment(e);
     }
+    /**
+     * Switches to UpdateAppointment Scene
+     * also sets static variable selectedAppointment that can be accessed from new scene
+     * @param e
+     * @throws IOException
+     */
     public void onUpdateAppointment(ActionEvent e) throws IOException {
         selectedAppointment = (Appointment)appointmentsTable.getSelectionModel().getSelectedItem();
         if(selectedAppointment == null){
@@ -109,6 +153,12 @@ public class MainMenuController implements Initializable{
         }
 
     }
+    /**
+     * Switches to UpdateCustomer scene
+     * Also sets static variable selectedCusomer that can be accessed from new scene
+     * @param e
+     * @throws IOException
+     */
     public void onUpdateCustomer(ActionEvent e) throws IOException {
         selectedCustomer = (Customer)customersTable.getSelectionModel().getSelectedItem();
         if(selectedCustomer == null){
@@ -118,6 +168,11 @@ public class MainMenuController implements Initializable{
             Navigation.switchToUpdateCustomer(e);
         }
     }
+    /**
+     * if cusomter is selected and has no appointments, Pop up alert double checks
+     * If user accepts, customer is deleted. Informative errors for all cases
+     * @throws SQLException
+     */
     public void onDeleteCustomer() throws SQLException {
         int selectedID;
         Customer selected;
@@ -129,17 +184,27 @@ public class MainMenuController implements Initializable{
             errorLabel.setText("No Customer Selected!");
             return;
         }
-        if(AppointmentsQuery.getCustomerAppointments(selectedID).isEmpty()){
+        if(AppointmentsQuery.customerFilter.extract(AppointmentsQuery.getAllAppointments(),selectedID).isEmpty()){
             String name = selected.getCustomer_Name();
-            CustomersQuery.delete(selectedID);
-            customersTable.setItems(CustomersQuery.getCustomers());
-            appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
-            errorLabel.setText("Customer: " + name + " has been deleted!");
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION,"Are you sure you want to Delete: " + name + "?");
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                CustomersQuery.delete(selectedID);
+                customersTable.setItems(CustomersQuery.getCustomers());
+                appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
+                errorLabel.setText("Customer: " + name + " has been deleted!");
+            }
+            else errorLabel.setText("");
         }
         else{
             errorLabel.setText("Cannot Delete a Customer with Appointments!");
         }
     }
+    /**
+     * if appointment is selected, Pop up alert double checks
+     * If user accepts, appointment is deleted. Informative errors for all cases
+     * @throws SQLException
+     */
     public void onDeleteAppointment() throws SQLException {
         int selectedID;
         Appointment selected;
@@ -152,19 +217,48 @@ public class MainMenuController implements Initializable{
             return;
         }
         String title = selected.getTitle();
-        AppointmentsQuery.delete(selectedID);
-        appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
-        errorLabel.setText("Appointment: " + title + " has been deleted!");
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION,"Are you sure you want to Delete: " + title + "?");
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            AppointmentsQuery.delete(selectedID);
+            appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
+            errorLabel.setText("Appointment: " + title + " has been deleted!");
+        }
+        else errorLabel.setText("");
     }
+
+    /**
+     * Switches to Reporting Scene
+     * @param e
+     * @throws IOException
+     */
+    public void onReporting(ActionEvent e) throws IOException {
+        Navigation.switchToReporting(e);
+    }
+
+    /**
+     * Takes user back to login screen
+     * @param e
+     * @throws IOException
+     */
     public void onLogOut(ActionEvent e) throws IOException {
         Navigation.logOut(e);
     }
+
+    /**
+     * Initialize method to populate tables, check for urgent appointment
+     * Lambda expression defined in DAO class is used to filter appointments based on start time
+     * in order to check for urgent appointmnets
+     * @param url
+     * @param resourceBundle
+     */
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle){
         welcomeLabel.setText("Welcome, " + LoginController.username + "!");
         //CHECK FOR URGENT APPOINTMENTS
         try {
-            ObservableList<Appointment> urgentAppointments = AppointmentsQuery.getUrgentAppointments();
+            ObservableList<Appointment> urgentAppointments =
+                    AppointmentsQuery.urgentFilter.extract(AppointmentsQuery.getAllAppointments(),15);
             if(urgentAppointments.isEmpty()){
                 urgentAppointmentLabel.setText("No Urgent Appointments!");
                 System.out.println("No urgent appt detected");
@@ -182,6 +276,7 @@ public class MainMenuController implements Initializable{
         //Update Customer List upon Main Menu Load
         try {
             customersTable.setItems(CustomersQuery.getCustomers());
+            appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
@@ -192,11 +287,6 @@ public class MainMenuController implements Initializable{
         customersDivisionIDColumn.setCellValueFactory(new PropertyValueFactory<>("division_ID"));
         customersPhoneColumn.setCellValueFactory(new PropertyValueFactory<>("phone"));
 
-        try {
-            appointmentsTable.setItems(AppointmentsQuery.getAllAppointments());
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
         appointmentsIDColumn.setCellValueFactory(new PropertyValueFactory<>("appointment_ID"));
         appointmentsTitleColumn.setCellValueFactory(new PropertyValueFactory<>("title"));
         appointmentsDescriptionColumn.setCellValueFactory(new PropertyValueFactory<>("description"));
